Add a way to store constants
Our chunks can now manage constants for our programs using a dynamic
value_array. The logic behind this array is the same as that of the
chunk opcodes. We use the same functions to manage memory when growing
the array. Similar to the chunks, this gives us on average O(1) cost
when reallocating the array.